phase 3 module refactoring
