+++ native/src/batch_retrieval/simple.rs
